Move the board memory fingerprint into component_manager
Both checksum paths already reach component_manager, so the helper does not
need a module of its own. arduino.py imports it directly and espidf.py takes it
from the module it already loads, which drops the extra file and the loader
block that came with it.
